Updating no-NB.json
Updated no-NB.json with new entries, some changes and typos fixed.

Some feedbacks to the translation:
I see there is some missing entries like title Language, Open Langauge Editor in Language tab in Settings. Also under Manage Shortcuts, it's missing entries for Shortcut and Date added in Shortcuts tab in Settings, it could be maybe it's just in English despite for it has new entries for Norwegian. 
My Games entry looks to be not working despite it's correct written because it just appears to be in English that is a bug.

About the DELETE command to delete account, I am not sure if it should be translated or not. I just left the English word command for now. If yes, it should be SLETT command since this is correct word in Norwegian.